refacto: tests
