+++ test/test_datasets_loaders.py
@@ -1,6 +1,23 @@
 from doctr.datasets import loaders
 from typing import Tuple
 import tensorflow as tf
+import pytest
+import json
+
+
+@pytest.fixture(scope="session")
+def mock_detection_label(tmpdir_factory):
+    folder = tmpdir_factory.mktemp("labels")
+    label = {
+        "boxes_1": [[[1, 2], [1, 3], [2, 1], [2, 3]], [[10, 20], [10, 30], [20, 10], [20, 30]]],
+        "boxes_2": [[[3, 2], [3, 3], [4, 1], [4, 3]], [[30, 20], [30, 30], [40, 10], [40, 30]]],
+        "boxes_3": [[[1, 5], [1, 8], [2, 5], [2, 8]]],
+    }
+    for i in range(5):
+        fn = folder.join("mock_image_file_" + str(i) + ".jpeg.json")
+        with open(fn, 'w') as f:
+            json.dump(label, f)
+    return str(folder)
 
 
 def test_detection_core_generator(mock_image_folder, mock_detection_label):
@@ -14,5 +31,6 @@ def test_detection_core_generator(mock_image_folder, mock_detection_label):
     for _, batch in enumerate(core_loader):
         image, gt, mask = batch
         assert isinstance(image, tf.Tensor)
+        assert image.shape[1] == image.shape[2] == 1024
         assert isinstance(gt, tf.Tensor)
         assert isinstance(mask, tf.Tensor)
